pyproject.toml
[project]
name = "finance-mcp-server"
version = "0.1.0"
description = "MCP server exposing Yahoo Finance stock price tools via FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "fastmcp>=3.1.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.1.1",
    "rich>=14.3.3",
    "scalekit-sdk-python>=2.7.1",
    "tenacity>=9.1.4",
    "yfinance>=1.2.0",
]

[project.scripts]
finance-mcp-server = "finance_mcp_server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=9.11.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "tools: MCP tool tests",
    "middleware: auth middleware tests",
    "routes: HTTP route tests",
    "config: configuration tests",
]

[tool.ruff]
line-length = 130
target-version = "py312"
preview = true

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
    "SIM",
    "TRY",
    "PERF",
    "T20",
    "ERA",
    "ASYNC",
    "RUF",
    "DOC",
    "D",
]

[tool.ruff.lint.pydocstyle]
convention = "google"
